Move models to feature tier
models owns the /model/* route, so it's a feature, not shared. Flip its
tier in module-boundaries.mjs and resolve the two violations:

- FormModelPicker: a self-contained model picker (no models-internal deps)
  used only by actions/ActionCreator/CreateActionForm — moved into actions
  next to its sole consumer.
- NewModelOptions: the /model/new page, referenced by the query_builder
  container test harness (test-utils.tsx) to wire the route. test-utils is
  not boundary-exempt; injected the component via a new optional
  `newModelOptionsComponent` param (defaulting to a stub) and pass the real
  NewModelOptions from the two specs that render /model/new — spec files
  are exempt from boundary rules.

No behavior change.

DEV-2133
