+++ frontend/src/metabase/query_builder/containers/test-utils.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable i18next/no-literal-string */
 import userEvent from "@testing-library/user-event";
 import fetchMock from "fetch-mock";
-import type { ComponentPropsWithoutRef } from "react";
+import type { ComponentPropsWithoutRef, ComponentType } from "react";
 import { IndexRoute, Route } from "react-router";
 
 import {
@@ -36,7 +36,6 @@ import {
 import { NewItemMenu } from "metabase/common/components/NewItemMenu";
 import { LOAD_COMPLETE_FAVICON } from "metabase/common/hooks/constants";
 import { serializeCardForUrl } from "metabase/common/utils/card";
-import NewModelOptions from "metabase/models/containers/NewModelOptions";
 import { createMockState } from "metabase/redux/store/mocks";
 import { checkNotNull } from "metabase/utils/types";
 import type { Card, Dataset, UnsavedCard } from "metabase-types/api";
@@ -219,6 +218,12 @@ const TestQueryBuilder = (
 
 const TestHome = () => <NewItemMenu trigger={<button>New</button>} />;
 
+// The real /model/new page (NewModelOptions) lives in the `models` feature.
+// Feature modules can't import each other, so tests that need it inject it via
+// `newModelOptionsComponent` (spec files are exempt from boundary rules); other
+// tests never hit this route and get this stub.
+const TestNewModelOptions = () => <div />;
+
 const TestRedirect = () => <div />;
 
 const isSavedCard = (card: Card | UnsavedCard | null): card is Card => {
@@ -229,6 +234,9 @@ interface SetupOpts {
   card: Card | UnsavedCard | null;
   dataset?: Dataset;
   initialRoute?: string;
+  // Loosely typed to match react-router's `component` prop: the real
+  // NewModelOptions receives router-injected props (location) the stub omits.
+  newModelOptionsComponent?: ComponentType<any>;
 }
 
 export const setup = async ({
@@ -241,6 +249,7 @@ export const setup = async ({
         ? `/${card.id}`
         : `#${serializeCardForUrl(card)}`
   }`,
+  newModelOptionsComponent = TestNewModelOptions,
 }: SetupOpts) => {
   setupUserMetabotPermissionsEndpoint();
   setupDatabasesEndpoints([TEST_DB]);
@@ -284,7 +293,7 @@ export const setup = async ({
       <Route>
         <Route path="/" component={TestHome} />
         <Route path="/model">
-          <Route path="new" component={NewModelOptions} />
+          <Route path="new" component={newModelOptionsComponent} />
           <Route path="query" component={TestQueryBuilder} />
           <Route path="columns" component={TestQueryBuilder} />
           <Route path="metadata" component={TestQueryBuilder} />
